+++ frontend/todofrontend/src/app/uncomplete/uncomplete.component.ts
@@ -9,37 +9,59 @@ import { TodoserviceService } from '../services/todoservice.service';
 export class UncompleteComponent {
 
   unDoneList = []
-
-
   
   checkUndone(){
-    this.unDoneList= this.todo.getUserData().filter(item=>{
-
+    this.unDoneList= this.todo.getUserData().filter((item: any)=>{
       return item['completed'] == false
     })
 }
 
-  handleDelete(id : string ){
-    this.todo.deleteATodo(id).subscribe(res=>{
-     alert('item successfully deleted')
-     // console.log(res)
-     this.todo.getUserList().subscribe(resp=>{
-       console.log(resp)
-       if(resp.todos.length == 0 ){
-         
-         this.todo.setUserData(resp.todos)
-         this.checkUndone()
- 
-       }else{
-         this.todo.setUserData(resp.todos)
-         this.checkUndone()
- 
-       }
-      })  
- 
+
+
+handleDelete(id: string) {
+
+  let choice = confirm("Are you sure you want to delete this item?")
+  if (choice){
+    this.todo.deleteATodo(id).subscribe(res => {
+      alert('item successfully deleted')
+      // console.log(res)
+      this.todo.getUserList().subscribe(resp => {
+        if (resp.todos.length == 0) {
+          this.todo.setUserData(resp.todos)
+          this.checkUndone()
+        } 
+          this.todo.setUserData(resp.todos)
+          this.checkUndone()
+
+      })
+
     })
+  }
+  return
+
+
+}
+  // handleDelete(id : string ){
+  //   this.todo.deleteATodo(id).subscribe(res=>{
+  //    alert('item successfully deleted')
+  //    // console.log(res)
+  //    this.todo.getUserList().subscribe(resp=>{
+  //      console.log(resp)
+  //      if(resp.todos.length == 0 ){
+         
+  //        this.todo.setUserData(resp.todos)
+  //        this.checkUndone()
+ 
+  //      }
+  //        this.todo.setUserData(resp.todos)
+  //        this.checkUndone()
+ 
+       
+  //     })  
+ 
+  //   })
     
-   }
+  //  }
    handleComplete(id:string){
     this.todo.completeAToDo(id).subscribe(res=>{
       this.todo.getUserList().subscribe(response=>{
@@ -51,6 +73,13 @@ export class UncompleteComponent {
   constructor(public todo : TodoserviceService){}
 
 ngOnInit(){
-  this.checkUndone()
+  this.todo.getUserList().subscribe({
+    next: res => {
+      this.todo.setUserData(res.todos)
+      this.checkUndone()
+    },
+    error: err => {
+    }
+  })
 }
 }
